+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Wang 2019 tacrolimus ([doi:10.3892/etm.2019.7446](https://doi.org/10.3892/etm.2019.7446)) -- Chinese children aged 2.4-16.4 years with refractory nephrotic syndrome.
+- Add Guidi 2019 artesunate and mefloquine ([doi:10.1186/s12936-019-2754-6](https://doi.org/10.1186/s12936-019-2754-6)) -- African children aged 6-59 months with uncomplicated falciparum malaria.
 
 - Add Palmer 2025 moxifloxacin ([doi:10.1002/bcp.70005](https://doi.org/10.1002/bcp.70005)) -- children with rifampicin-resistant tuberculosis.